Extract Pivot and Elevator classes
