+++ src/main/java/frc/robot/commands/arm/elevator/RunElevatorToExtensionCommand.java
@@ -10,9 +10,10 @@ import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Constants.ArmConstants.ElevatorConstants;
 import frc.robot.subsystems.Arm;
+import frc.robot.subsystems.arm.Elevator;
 
 public class RunElevatorToExtensionCommand extends CommandBase {
-  private final Arm.Elevator elevator;
+  private final Elevator elevator;
   private final DoubleSupplier extension;
 
   /**
